Implement Win32 condition variable
Due to a bug in C++11 cond_timedwait on Windows, this change adds a Win32 condition implementation that also works on Windows CE, XP, and Vista.  It uses a waiting thread counter as part of the condition object to ensure only a waiting thread is unblocked.  Since all code using this facility at this point only wait in one thread this change also changes the Post method for C11 and pthreads implementation to (pthreads)_cond_signal to make the implementation of the win32 condition simpler, although it would be possible to change the event into a manual reset event to have the same behavior as broadcast as the code takes care of resetting the event again when no more waiting threads are present to prevent a race condition.