Force use_openssl=ON in CMake when use_wsio=ON
The CMake option "use_wsio" pulls in libwebsockets. Libwebsockets uses OpenSSL
by default (it can use WolfSSL or CyaSSL instead, but I don't think we're
taking advantage of these options today). Libwebsockets does some funny stuff
in CMake, which I was careful to handle in a recent commit, in the case where
the user set use_openssl=ON.

Now we're seeing a problem in the case where the user turns use_wsio ON, but
leaves use_openssl OFF (e.g. on Windows--the Jenkins builds do this). I'm
fixing it by forcing use_openssl=ON in this particular case. Then
libwebsockets gets set up properly, and code can find the needed headers and
libs.