Remove SECURITY_FLAG_IGNORE_CERT_CN_INVALID flag.
Was needing this in local test because was setting 127.0.0.1 as target but the cert was to my machine name.

Correct fix however is for caller to make sure names align, which is point of SSL.

C SDK's use of Curl has same behavior on mismatched names, so as bonus this brings Windows/Linux parity.