Revert change of dns_resolver_create in socketio
The additonal change related to dns_resolver_create in socketio_* unveiled another set of issues related to the
handling of hostname in the socketio structure. To provide proper fixes, the changes will be split into
separate commits, first to address github #1275 and a subsequent one to address the remaining issues.
Thus reverting part of the recent changes to align with the strategy of having separate fixes.